- added \Yiisoft\Yii\Cycle\Data\Writer\EntityWriter - \Yiisoft\Yii\Cycle\DataReader\SelectDataReader moved to \Yiisoft\Yii\Cycle\Data\Reader\EntityReader
