+++ src/Data/Reader/Processor/GreaterThanOrEqual.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Yii\Cycle\DataReader\Processor;
+namespace Yiisoft\Yii\Cycle\Data\Reader\Processor;
 
 final class GreaterThanOrEqual extends CompareProcessor
 {
