+++ src/Data/Reader/Processor/In.php
@@ -2,15 +2,16 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Yii\Cycle\DataReader\Processor;
+namespace Yiisoft\Yii\Cycle\Data\Reader\Processor;
 
+use Spiral\Database\Injection\Parameter;
 use Yiisoft\Data\Reader\Filter\FilterProcessorInterface;
 
-final class Like implements QueryBuilderProcessor, FilterProcessorInterface
+final class In implements QueryBuilderProcessor, FilterProcessorInterface
 {
     public function getOperator(): string
     {
-        return 'like';
+        return 'in';
     }
 
     public function getAsWhereArguments(array $arguments, array $processors): array
@@ -21,6 +22,6 @@ final class Like implements QueryBuilderProcessor, FilterProcessorInterface
 
         [$field, $value] = $arguments;
 
-        return [$field, $this->getOperator(), $value];
+        return [$field, $this->getOperator(), new Parameter($value)];
     }
 }
