+++ src/Data/Reader/Processor/LessThanOrEqual.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Yii\Cycle\DataReader\Processor;
+namespace Yiisoft\Yii\Cycle\Data\Reader\Processor;
 
 final class LessThanOrEqual extends CompareProcessor
 {
