+++ src/Data/Reader/Processor/Like.php
@@ -2,16 +2,15 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Yii\Cycle\DataReader\Processor;
+namespace Yiisoft\Yii\Cycle\Data\Reader\Processor;
 
-use Spiral\Database\Injection\Parameter;
 use Yiisoft\Data\Reader\Filter\FilterProcessorInterface;
 
-final class In implements QueryBuilderProcessor, FilterProcessorInterface
+final class Like implements QueryBuilderProcessor, FilterProcessorInterface
 {
     public function getOperator(): string
     {
-        return 'in';
+        return 'like';
     }
 
     public function getAsWhereArguments(array $arguments, array $processors): array
@@ -22,6 +21,6 @@ final class In implements QueryBuilderProcessor, FilterProcessorInterface
 
         [$field, $value] = $arguments;
 
-        return [$field, $this->getOperator(), new Parameter($value)];
+        return [$field, $this->getOperator(), $value];
     }
 }
